Further package reorganisation
