+++ src/main/java/eu/europa/ted/eforms/sdk/entity/SdkFieldRepository.java
@@ -1,12 +1,10 @@
-package eu.europa.ted.eforms.sdk.metadata;
+package eu.europa.ted.eforms.sdk.entity;
 
 import java.nio.file.Path;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import eu.europa.ted.eforms.sdk.SdkConstants;
 import eu.europa.ted.eforms.sdk.helpers.MapFromJson;
-import eu.europa.ted.efx.model.EfxEntityFactory;
-import eu.europa.ted.efx.model.SdkField;
 
 public class SdkFieldRepository extends MapFromJson<SdkField> {
   private static final long serialVersionUID = 1L;
@@ -19,7 +17,7 @@ public class SdkFieldRepository extends MapFromJson<SdkField> {
   protected void populateMap(final JsonNode json) throws InstantiationException {
     final ArrayNode fields = (ArrayNode) json.get(SdkConstants.FIELDS_JSON_FIELDS_KEY);
     for (final JsonNode field : fields) {
-      final SdkField sdkField = EfxEntityFactory.getSdkField(sdkVersion, field);
+      final SdkField sdkField = SdkEntityFactory.getSdkField(sdkVersion, field);
       put(sdkField.getId(), sdkField);
     }
   }
