+++ src/main/java/eu/europa/ted/eforms/sdk/entity/SdkNodeRepository.java
@@ -1,12 +1,10 @@
-package eu.europa.ted.eforms.sdk.metadata;
+package eu.europa.ted.eforms.sdk.entity;
 
 import java.nio.file.Path;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ArrayNode;
 import eu.europa.ted.eforms.sdk.SdkConstants;
 import eu.europa.ted.eforms.sdk.helpers.MapFromJson;
-import eu.europa.ted.efx.model.EfxEntityFactory;
-import eu.europa.ted.efx.model.SdkNode;
 
 public class SdkNodeRepository extends MapFromJson<SdkNode> {
   private static final long serialVersionUID = 1L;
@@ -19,7 +17,7 @@ public class SdkNodeRepository extends MapFromJson<SdkNode> {
   protected void populateMap(final JsonNode json) throws InstantiationException {
     final ArrayNode nodes = (ArrayNode) json.get(SdkConstants.FIELDS_JSON_XML_STRUCTURE_KEY);
     for (final JsonNode node : nodes) {
-      final SdkNode sdkNode = EfxEntityFactory.getSdkNode(sdkVersion, node);
+      final SdkNode sdkNode = SdkEntityFactory.getSdkNode(sdkVersion, node);
       put(sdkNode.getId(), sdkNode);
     }
   }
